ci: Update GitHub Actions workflows for container publishing
Refactor the GitHub Actions workflows to update the triggers for container publishing. Change the trigger for the "build_and_test.yml" workflow to only run on pull requests to the main branch. Additionally, add a new workflow "publish_containers.yaml" to handle the publishing of containers. This workflow will be triggered on pushes to the dev branch.

Related recent commits:
- ci: change upstream registry
- ci: temporarily add listing of directory contents for debugging purposes
- ci: add step in container build/test workflow to download already existing containers